Fix AM_PATH_GTK_2_0 for cross compilation
With the current code used in the m4 macro, gtk-2.0.m4 can't find
a prefixed pkg-config. This presents issue when cross compiling, as
pkg-config needs to be prefixed to work with the target being compiled
for.

Change the macro to use PKG_PROG_PKG_CONFIG rather than custom
detection code.

https://bugzilla.gnome.org/show_bug.cgi?id=749507