placessidebar: fix Connect to Server position
Commit 7db399d975b8f9626c21761dde5f2d5feeb6e305 introduces
the Other Locations item, as well as some behavioral changes
to GtkPlacesSidebar. It, howevers, wrongly changes the section
of code that handles the Connect to Server item, with various
side effects depending on the environment setup.

Fix that by adding the Connect to Server at the right section
of code.